feat: 수익률 계산 로직 구현

+++ src/main/java/lotto/view/OutputView.java
@@ -17,7 +17,7 @@ public class OutputView {
 
     }
 
-    public void printProfit(float profit) {
+    public void printProfit(Double profit) {
 
     }
